[refact]: SW Architecture에 맞게 디렉터리 구조 수정
