+++ situationDetector/README.md
@@ -1,0 +1,22 @@
+# AI_Server
+
+## 1. 개요
+
+## 2. 통신
+- 데이터 수신
+  1. PED로부터 영상 메타데이터 수신 (TCP, json 형식, 1초마다 수신) (PED ->)
+  2. TCP 연결이 수립됩 동안, PED로부터 영상 픽셀데이터 수신 (UDP) (PED ->)
+
+- 데이터 발신
+  1. YOLO로 분석한 결과를 DB 서버에 발신함 (-> DB Server)
+  2. 사건 전후 15초 영상을 DB 서버에 발신함 (-> DB Server)
+  3. 받아온 영상 픽셀데이터와 메타데이터를 그대로 GUI에 발신함 (-> GUI)
+
+## 3. 작동 예시
+
+
+## 4. 구조
+- ai_server.py : 스레드 생성 및 관리 등의 역활을 수행하는 서버의 main 역활 파일
+    1. tcp_ped_receiver.py : 순찰차 영상의 메타데이터를 1초마다 받아오는 작업을 하는 함수 파일
+    2. udp_ped_receiver.py : 순찰차 영상의 픽셀데이터를 받아오는 작업을 하는 함수 파일
+    3. yolo_
