+++ situationDetector/tcp_db_sender.py
@@ -1,0 +1,11 @@
+# tcp_sender.py
+import socket
+import json
+import time
+import geocoder
+import threading
+
+# 통신 변수
+TCP_HOST = 'localhost'
+TCP_PORT = 6602
+
