+++ situationDetector/tcp_ped_receiver.py
@@ -1,0 +1,52 @@
+# tcp_receiver.py
+"""
+PED로부터 영상 메타데이터를 수신하는 클라이언트
+(PED) -> (AI_Server)
+"""
+import socket
+import json
+import threading
+
+TCP_HOST = '0.0.0.0'  # 모든 인터페이스에서 연결 허용
+TCP_PORT = 6600
+
+def handle_tcp_client(shutdown_event: threading.Event):
+    """
+    TCP 클라이언트의 연결을 수락하고 메타데이터를 수신
+    """
+    server_sock = socket.socket(socket.AF_INET, socket.SOCK_STREAM)
+    # SO_REUSEADDR 옵션을 설정하여 주소 재사용 문제를 방지
+    server_sock.setsockopt(socket.SOL_SOCKET, socket.SO_REUSEADDR, 1)
+    server_sock.bind((TCP_HOST, TCP_PORT))
+    server_sock.listen()
+    # accept() 호출이 블로킹되지 않도록 타임아웃 설정
+    server_sock.settimeout(1.0)
+    print(f"AI Server (TCP) : 서버가 {TCP_HOST}:{TCP_PORT}에서 연결 대기")
+
+    while not shutdown_event.is_set():
+        try:
+            conn, addr = server_sock.accept()
+            with conn:
+                print(f"AI Server (TCP) : 클라이언트 연결됨: {addr}")
+                while not shutdown_event.is_set():
+                    # 데이터 수신 (4096 바이트 버퍼)
+                    data = conn.recv(4096)
+                    if not data:
+                        print(f"AI Server (TCP) : 클라이언트 연결 끊어짐: {addr}")
+                        break
+                    
+                    try:
+                        metadata = json.loads(data.decode('utf-8'))
+                        print(f"AI Server (TCP) : 수신된 메타데이터: {metadata}")
+                    except (json.JSONDecodeError, UnicodeDecodeError) as e:
+                        print(f"AI Server (TCP) : 데이터 파싱 오류: {e}")
+
+        except socket.timeout:
+            # 타임아웃은 정상적인 상황이므로 루프를 계속 진행
+            continue
+        except Exception as e:
+            print(f"AI Server (TCP) : 서버 오류: {e}")
+            break
+    
+    server_sock.close()
+    print("AI Server (TCP) : 수신 스레드를 종료합니다.")
